use workspace setting to determine error grouping
