+++ backend/model/model.go
@@ -379,9 +379,10 @@ type ProjectFilterSettings struct {
 
 type AllWorkspaceSettings struct {
 	Model
-	WorkspaceID   int  `gorm:"uniqueIndex"`
-	AIApplication bool `gorm:"default:true"`
-	AIInsights    bool `gorm:"default:false"`
+	WorkspaceID     int  `gorm:"uniqueIndex"`
+	AIApplication   bool `gorm:"default:true"`
+	AIInsights      bool `gorm:"default:false"`
+	ErrorEmbeddings bool `gorm:"default:false"`
 }
 
 type HasSecret interface {
